All apis work correctly without /pin/new
