+++ lib/api/boards_api.dart
@@ -1,52 +1,10 @@
 import 'dart:convert';
 
-import 'package:flutter/cupertino.dart';
 import 'package:mobile/api/api_client.dart';
 import 'package:mobile/model/board_model.dart';
 import 'package:mobile/model/pin_model.dart';
 
-class NewBoard {
-  NewBoard({
-    @required this.name,
-    this.description,
-    @required this.isPrivate,
-  });
-
-  String name;
-  String description;
-  bool isPrivate;
-
-  Map<String, dynamic> toJson() => {
-        'name': name,
-        'description': description ?? '',
-        'isPrivate': isPrivate,
-      };
-}
-
-class EditBoard {
-  EditBoard({
-    this.name,
-    this.description,
-    this.isPrivate,
-    this.isArchive,
-  });
-
-  String name;
-  String description;
-  bool isPrivate;
-  bool isArchive;
-
-  Map<String, dynamic> toJson() => {
-        'name': name,
-        'description': description,
-        'isPrivate': isPrivate,
-        'isArchive': isArchive
-      };
-}
-
 abstract class BoardsApi {
-  Future<List<Board>> boards();
-
   Future<Board> newBoard({NewBoard board});
 
   Future<bool> deleteBoard({int id});
@@ -66,14 +24,6 @@ class DefaultBoardsApi extends BoardsApi {
     final response = await _client.get("/boards/$id/pins?page=$page");
     return (jsonDecode(response.body) as List)
         .map((it) => Pin.fromJson(it))
-        .toList();
-  }
-
-  @override
-  Future<List<Board>> boards() async {
-    final response = await _client.get("/boards");
-    return (jsonDecode(response.body) as List)
-        .map((it) => Board.fromJson(it))
         .toList();
   }
