+++ lib/model/board_model.dart
@@ -1,3 +1,47 @@
+import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+part 'board_model.g.dart';
+
+@JsonSerializable(includeIfNull: false)
+class NewBoard {
+  NewBoard({
+    @required this.name,
+    this.description,
+    @required this.isPrivate,
+  });
+
+  String name;
+  String description;
+  bool isPrivate;
+
+  factory NewBoard.fromJson(Map<String, dynamic> json) =>
+      _$NewBoardFromJson(json);
+
+  Map<String, dynamic> toJson() => _$NewBoardToJson(this);
+}
+
+@JsonSerializable(includeIfNull: false)
+class EditBoard {
+  EditBoard({
+    this.name,
+    this.description,
+    this.isPrivate,
+    this.isArchive,
+  });
+
+  String name;
+  String description;
+  bool isPrivate;
+  bool isArchive;
+
+  factory EditBoard.fromJson(Map<String, dynamic> json) =>
+      _$EditBoardFromJson(json);
+
+  Map<String, dynamic> toJson() => _$EditBoardToJson(this);
+}
+
+@JsonSerializable(includeIfNull: false)
 class Board {
   final int id;
   final int userId;
@@ -15,13 +59,9 @@ class Board {
     this.isArchive,
   });
 
-  Board.fromJson(Map<String, dynamic> json)
-      : id = json['id'] as int,
-        userId = json['userId'] as int,
-        name = json['name'],
-        description = json['description'],
-        isPrivate = json['isPrivate'],
-        isArchive = json['isArchive'];
+  factory Board.fromJson(Map<String, dynamic> json) => _$BoardFromJson(json);
+
+  Map<String, dynamic> toJson() => _$BoardToJson(this);
 
   static Board fromMock() {
     return Board(
