+++ lib/model/pin_model.dart
@@ -1,3 +1,48 @@
+import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+part 'pin_model.g.dart';
+
+@JsonSerializable(includeIfNull: false)
+class NewPin {
+  NewPin({
+    @required this.title,
+    @required this.image,
+    this.description,
+    this.url,
+    this.isPrivate,
+  });
+
+  final String title;
+  final String description;
+  final String url;
+  final bool isPrivate;
+  final String image;
+
+  factory NewPin.fromJson(Map<String, dynamic> json) => _$NewPinFromJson(json);
+
+  Map<String, dynamic> toJson() => _$NewPinToJson(this);
+}
+
+@JsonSerializable(includeIfNull: false)
+class EditPin {
+  EditPin({
+    this.title,
+    this.description,
+    this.isPrivate,
+  });
+
+  final String title;
+  final String description;
+  final bool isPrivate;
+
+  factory EditPin.fromJson(Map<String, dynamic> json) =>
+      _$EditPinFromJson(json);
+
+  Map<String, dynamic> toJson() => _$EditPinToJson(this);
+}
+
+@JsonSerializable(includeIfNull: false)
 class Pin {
   final int id;
   final String title;
@@ -17,14 +62,9 @@ class Pin {
     this.isPrivate,
   });
 
-  Pin.fromJson(Map<String, dynamic> json)
-      : id = json['id'] as int,
-        title = json['title'],
-        description = json['description'],
-        url = json['url'],
-        userId = json['userId'] as int,
-        imageUrl = json['imageUrl'],
-        isPrivate = json['isPrivate'] as bool;
+  factory Pin.fromJson(Map<String, dynamic> json) => _$PinFromJson(json);
+
+  Map<String, dynamic> toJson() => _$PinToJson(this);
 
   static Pin fromMock() {
     return Pin(
